Refactoring
- Operations refactoring to separate files
- save GCD refactoring
